--- a/makefile
+++ b/makefile
@@ -1,8 +1,8 @@
 build:
 	py -m build
 
-install: dist/multi_bible_search-2.1.2.tar.gz
-	pip install --force-reinstall ./dist/multi_bible_search-2.1.2.tar.gz
+install: dist/multi_bible_search-2.1.3.tar.gz
+	pip install --force-reinstall ./dist/multi_bible_search-2.1.3.tar.gz
 	copy venv\\Lib\\site-packages\\multi_bible_search\\*.pyd src\\multi_bible_search\\
 
 full: build install