--- a/pyproject.toml
+++ b/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "multi_bible_search"
-version = "2.1.2"
+version = "2.1.3"
 authors = [
   { name="Samuel Howard" },
 ]