### Improvements:

- Reworked how indices are stored before being ranked to improve worst-case performance

- Various other C code optimizations

#### Before:

- Total object size of 60.8 MiB

- Single-loaded version (KJV) object size of 7.2 MiB

- ~54μs per query (on my machine)

- ~1.1s per excessively long query (on my machine)

- Index size of 18.1 MiB

#### Now:

- Total object size of 60.8 MiB

- Single-loaded version (KJV) object size of 7.2 MiB

- ~48μs per query (on my machine)

- ~0.02411s per excessively long query (on my machine)

- Index size of 18.1 MiB